// app/types.ts
export type RootStackParamList = {
    PersonalTrackerMain : undefined;
    PersonalTrackerIncome: undefined;
    PersonalTrackerExpense: undefined;
    TransactionHistory : undefined;
    commonregistration1 : {
      username: string;
      phoneNumber: string;
      address: string;
      district: string;
    }; 
    commonregistration2 : {
      username: string;
      phoneNumber: string;
      address: string;
      district: string;
      Email : " ", 
      Password: " "
    };
    commonregistration3 : {
      username: string;
      phoneNumber: string;
      address: string;
      district: string;
      Email: string;
      Password: string;
    };
    login : undefined;
    Homepage : undefined;
    DiseaseIdentification: undefined;
    Fertilizerseller: undefined; 
    Weather : undefined;
    product : undefined;
    SelectLanguage : undefined;
    Chooserole : undefined;
    Buildyourfarmland : {username : string};
    MarketPrice1 : undefined; 
    MarketPrice2 :undefined;
    MarketPrice3 : {cropName : String}; 
    SupplementReminder1 : undefined; 
    SupplementReminder2 : undefined; 
    passwordReset : {email : string};
    DiseaseIdentification2 : undefined; 
  };
  